src/services/__tests__/episodes.service.test.ts
import { LoaderFunctionArgs } from "react-router-dom";
import getEpisodesByPodcastId from "../episodes.service";
import { Http } from "../http";

describe("Episodes Service", () => {
  let spy: jest.SpyInstance;
  beforeEach(() => {
    jest.clearAllMocks();
    localStorage.clear();
    spy = jest.spyOn(Http, "get").mockResolvedValue(mockResponse);
  });

  test("should return a list of episodes from the red when dont exist cache", async () => {
    const params: Partial<LoaderFunctionArgs> = {
      params: {
        id: "2",
      },
    };
    const expectedResponse = expectedData;

    const response = await getEpisodesByPodcastId(params as LoaderFunctionArgs);

    expect(expectedResponse).toEqual(response);
    expect(spy).toBeCalledTimes(1);
  });

  test("should return a list of episodes from the cache when exist data without expire", async () => {
    const params: Partial<LoaderFunctionArgs> = {
      params: {
        id: "2",
      },
    };
    const expectedResponse = expectedData;

    await getEpisodesByPodcastId(params as LoaderFunctionArgs);
    const response = await getEpisodesByPodcastId(params as LoaderFunctionArgs);

    expect(expectedResponse).toEqual(response);
    expect(spy).toBeCalledTimes(1);
  });
});

export const expectedData = {
  episodes: [
    {
      title: "Finneas",
      duration: "61:19",
      date: "3/30/2023",
      id: 1000606556604,
      episodeUrl:
        "https://www.podtrac.com/pts/redirect.mp3/pdst.fm/e/chtbl.com/track/42D75/traffic.megaphone.fm/CAD3340510551.mp3?updated=1680127825",
      description: "Finneas music education and evolution",
    },
    {
      title: "Rick Rubin",
      duration: "39:11",
      date: "3/30/2023",
      id: 1000606556393,
      episodeUrl:
        "https://www.podtrac.com/pts/redirect.mp3/pdst.fm/e/chtbl.com/track/42D75/traffic.megaphone.fm/CAD3099037887.mp3?updated=1680127893",
      description: "Rick Rubins music education and evolution",
    },
    {
      title: "Introducing This Little Light with Flea",
      duration: "1:30",
      date: "3/21/2023",
      id: 1000605224134,
      episodeUrl:
        "https://www.podtrac.com/pts/redirect.mp3/pdst.fm/e/chtbl.com/track/42D75/traffic.megaphone.fm/CAD8556787702.mp3?updated=1679428256",
      description: "Hosted by Flea, founding member",
    },
  ],
};

const mockResponse = {
  resultCount: 4,
  results: [
    {
      wrapperType: "track",
      kind: "podcast",
      collectionId: 1678378239,
      trackId: 1678378239,
      artistName: "Cadence13 and Parallel",
      collectionName: "This Little Light",
      trackName: "This Little Light",
      collectionCensoredName: "This Little Light",
      trackCensoredName: "This Little Light",
      collectionViewUrl:
        "https://podcasts.apple.com/us/podcast/this-little-light/id1678378239?uo=4",
      feedUrl: "https://feeds.megaphone.fm/this-little-light",
      trackViewUrl:
        "https://podcasts.apple.com/us/podcast/this-little-light/id1678378239?uo=4",
      artworkUrl30:
        "https://is1-ssl.mzstatic.com/image/thumb/Podcasts116/v4/f6/cf/66/f6cf66f5-8696-1903-f3b3-e95c507c8e9c/mza_5620205623101925377.jpg/30x30bb.jpg",
      artworkUrl60:
        "https://is1-ssl.mzstatic.com/image/thumb/Podcasts116/v4/f6/cf/66/f6cf66f5-8696-1903-f3b3-e95c507c8e9c/mza_5620205623101925377.jpg/60x60bb.jpg",
      artworkUrl100:
        "https://is1-ssl.mzstatic.com/image/thumb/Podcasts116/v4/f6/cf/66/f6cf66f5-8696-1903-f3b3-e95c507c8e9c/mza_5620205623101925377.jpg/100x100bb.jpg",
      collectionPrice: 0.0,
      trackPrice: 0.0,
      collectionHdPrice: 0,
      releaseDate: "2023-03-21T19:45:00Z",
      collectionExplicitness: "notExplicit",
      trackExplicitness: "explicit",
      trackCount: 3,
      trackTimeMillis: 90,
      country: "USA",
      currency: "USD",
      primaryGenreName: "Music Interviews",
      contentAdvisoryRating: "Explicit",
      artworkUrl600:
        "https://is1-ssl.mzstatic.com/image/thumb/Podcasts116/v4/f6/cf/66/f6cf66f5-8696-1903-f3b3-e95c507c8e9c/mza_5620205623101925377.jpg/600x600bb.jpg",
      genreIds: ["1525", "26", "1310"],
      genres: ["Music Interviews", "Podcasts", "Music"],
    },
    {
      episodeUrl:
        "https://www.podtrac.com/pts/redirect.mp3/pdst.fm/e/chtbl.com/track/42D75/traffic.megaphone.fm/CAD3340510551.mp3?updated=1680127825",
      previewUrl:
        "https://www.podtrac.com/pts/redirect.mp3/pdst.fm/e/chtbl.com/track/42D75/traffic.megaphone.fm/CAD3340510551.mp3?updated=1680127825",
      artistIds: [],
      feedUrl: "https://feeds.megaphone.fm/this-little-light",
      closedCaptioning: "none",
      collectionId: 1678378239,
      collectionName: "This Little Light",
      genres: [{ name: "Music Interviews", id: "1525" }],
      episodeGuid: "1726d660-ca86-11ed-bc57-df468bf305c7",
      description: "Finneas music education and evolution",
      trackId: 1000606556604,
      trackName: "Finneas",
      shortDescription: "",
      releaseDate: "2023-03-30T07:05:00Z",
      trackTimeMillis: 3679000,
      contentAdvisoryRating: "Explicit",
      trackViewUrl:
        "https://podcasts.apple.com/us/podcast/finneas/id1678378239?i=1000606556604&uo=4",
      artworkUrl60:
        "https://is1-ssl.mzstatic.com/image/thumb/Podcasts116/v4/f6/cf/66/f6cf66f5-8696-1903-f3b3-e95c507c8e9c/mza_5620205623101925377.jpg/60x60bb.jpg",
      collectionViewUrl:
        "https://itunes.apple.com/us/podcast/this-little-light/id1678378239?mt=2&uo=4",
      kind: "podcast-episode",
      wrapperType: "podcastEpisode",
      artworkUrl160:
        "https://is1-ssl.mzstatic.com/image/thumb/Podcasts116/v4/f6/cf/66/f6cf66f5-8696-1903-f3b3-e95c507c8e9c/mza_5620205623101925377.jpg/160x160bb.jpg",
      episodeFileExtension: "mp3",
      episodeContentType: "audio",
      artworkUrl600:
        "https://is1-ssl.mzstatic.com/image/thumb/Podcasts116/v4/f6/cf/66/f6cf66f5-8696-1903-f3b3-e95c507c8e9c/mza_5620205623101925377.jpg/600x600bb.jpg",
      country: "USA",
    },
    {
      episodeUrl:
        "https://www.podtrac.com/pts/redirect.mp3/pdst.fm/e/chtbl.com/track/42D75/traffic.megaphone.fm/CAD3099037887.mp3?updated=1680127893",
      previewUrl:
        "https://www.podtrac.com/pts/redirect.mp3/pdst.fm/e/chtbl.com/track/42D75/traffic.megaphone.fm/CAD3099037887.mp3?updated=1680127893",
      artistIds: [],
      feedUrl: "https://feeds.megaphone.fm/this-little-light",
      closedCaptioning: "none",
      collectionId: 1678378239,
      collectionName: "This Little Light",
      genres: [{ name: "Music Interviews", id: "1525" }],
      episodeGuid: "1819a5f2-ca86-11ed-bc57-2766a155c576",
      description: "Rick Rubins music education and evolution",
      trackId: 1000606556393,
      trackName: "Rick Rubin",
      shortDescription: "",
      releaseDate: "2023-03-30T07:01:00Z",
      trackTimeMillis: 2351000,
      contentAdvisoryRating: "Explicit",
      trackViewUrl:
        "https://podcasts.apple.com/us/podcast/rick-rubin/id1678378239?i=1000606556393&uo=4",
      artworkUrl60:
        "https://is1-ssl.mzstatic.com/image/thumb/Podcasts116/v4/f6/cf/66/f6cf66f5-8696-1903-f3b3-e95c507c8e9c/mza_5620205623101925377.jpg/60x60bb.jpg",
      collectionViewUrl:
        "https://itunes.apple.com/us/podcast/this-little-light/id1678378239?mt=2&uo=4",
      kind: "podcast-episode",
      wrapperType: "podcastEpisode",
      artworkUrl160:
        "https://is1-ssl.mzstatic.com/image/thumb/Podcasts116/v4/f6/cf/66/f6cf66f5-8696-1903-f3b3-e95c507c8e9c/mza_5620205623101925377.jpg/160x160bb.jpg",
      episodeFileExtension: "mp3",
      episodeContentType: "audio",
      artworkUrl600:
        "https://is1-ssl.mzstatic.com/image/thumb/Podcasts116/v4/f6/cf/66/f6cf66f5-8696-1903-f3b3-e95c507c8e9c/mza_5620205623101925377.jpg/600x600bb.jpg",
      country: "USA",
    },
    {
      episodeUrl:
        "https://www.podtrac.com/pts/redirect.mp3/pdst.fm/e/chtbl.com/track/42D75/traffic.megaphone.fm/CAD8556787702.mp3?updated=1679428256",
      previewUrl:
        "https://www.podtrac.com/pts/redirect.mp3/pdst.fm/e/chtbl.com/track/42D75/traffic.megaphone.fm/CAD8556787702.mp3?updated=1679428256",
      artistIds: [],
      feedUrl: "https://feeds.megaphone.fm/this-little-light",
      closedCaptioning: "none",
      collectionId: 1678378239,
      collectionName: "This Little Light",
      genres: [{ name: "Music Interviews", id: "1525" }],
      episodeGuid: "ff024fae-c820-11ed-bfc9-afb1f5412c09",
      description: "Hosted by Flea, founding member",
      trackId: 1000605224134,
      trackName: "Introducing This Little Light with Flea",
      shortDescription: "",
      releaseDate: "2023-03-21T19:45:00Z",
      trackTimeMillis: 90000,
      contentAdvisoryRating: "Explicit",
      trackViewUrl:
        "https://podcasts.apple.com/us/podcast/introducing-this-little-light-with-flea/id1678378239?i=1000605224134&uo=4",
      artworkUrl60:
        "https://is1-ssl.mzstatic.com/image/thumb/Podcasts116/v4/f6/cf/66/f6cf66f5-8696-1903-f3b3-e95c507c8e9c/mza_5620205623101925377.jpg/60x60bb.jpg",
      collectionViewUrl:
        "https://itunes.apple.com/us/podcast/this-little-light/id1678378239?mt=2&uo=4",
      kind: "podcast-episode",
      wrapperType: "podcastEpisode",
      artworkUrl160:
        "https://is1-ssl.mzstatic.com/image/thumb/Podcasts116/v4/f6/cf/66/f6cf66f5-8696-1903-f3b3-e95c507c8e9c/mza_5620205623101925377.jpg/160x160bb.jpg",
      episodeFileExtension: "mp3",
      episodeContentType: "audio",
      artworkUrl600:
        "https://is1-ssl.mzstatic.com/image/thumb/Podcasts116/v4/f6/cf/66/f6cf66f5-8696-1903-f3b3-e95c507c8e9c/mza_5620205623101925377.jpg/600x600bb.jpg",
      country: "USA",
    },
  ],
};
